Addressed more 64 bit compiler warnings.
  * misc/openvas_server.c (openvas_server_send): Cast argument to int
  when using it to specify field precision.

  * omp/xml.c (read_entity_and_string): Cast argument to int
  when using it to specify field precision.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@9423 423fd1db-d629-0410-8442-d21db03e70f4

+++ misc/openvas_server.c
@@ -508,7 +508,7 @@ openvas_server_send (gnutls_session_t * session, const char *string)
   while (left)
     {
       ssize_t count;
-      g_message ("   send %zu from %.*s[...]", left, left < 30 ? left : 30,
+      g_message ("   send %zu from %.*s[...]", left, left < 30 ? (int) left : 30,
                  string);
       count = gnutls_record_send (*session, string, left);
       if (count < 0)
@@ -541,7 +541,7 @@ openvas_server_send (gnutls_session_t * session, const char *string)
 
           return 1;
         }
-      g_message ("=> %.*s", count, string);
+      g_message ("=> %.*s", (int) count, string);
       string += count;
       left -= count;
     }
